# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "XR5_0TrainingRepo",
    "version": "1.0"
  },
  "paths": {
...
    "/xr50/training-repo/asset-management/Asset/{id}": {
      "get": {
        "tags": [
          "Asset"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Asset"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Asset"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Asset"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Asset"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Asset"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/xr50/training-repo/training-management/Training/{TrainingName}": {
      "get": {
        "tags": [
          "Training"
        ],
        "parameters": [
          {
            "name": "TrainingName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingModule"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingModule"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingModule"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Training"
        ],
        "parameters": [
          {
            "name": "TrainingName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingModule"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingModule"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingModule"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Training"
        ],
        "parameters": [
          {
            "name": "TrainingName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Asset": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "owncloudFileName": {
            "type": "string",
            "nullable": true
          },
          "trainingId": {
            "type": "string",
            "nullable": true
          },
          "resourceId": {
            "type": "string",
            "nullable": true
          },
          "assetId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ResourceBundle": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "owncloudFileName": {
            "type": "string",
            "nullable": true
          },
          "trainingId": {
            "type": "string",
            "nullable": true
          },
          "assetList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Asset"
            },
            "nullable": true,
            "readOnly": true
          },
          "resourceName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TrainingModule": {
        "type": "object",
        "properties": {
          "appName": {
            "type": "string",
            "nullable": true
          },
          "useCase": {
            "type": "string",
            "nullable": true
          },
          "resourceList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceBundle"
            },
            "nullable": true,
            "readOnly": true
          },
          "trainingName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}